+++ Omega2020/ai.py
@@ -9,17 +9,21 @@ import copy
 rows = 'ABCDEFGHI'
 cols = '123456789'
 
-boxes = [s+t for s in rows for t in cols]
-row_units = [[s+t for s in r for t in cols] for r in rows]
-column_units = [[s+t for s in rows for t in c] for c in cols]
-square_units = [[s+t for s in rs for t in cs]for rs in ('ABC','DEF','GHI') for cs in ('123','456','789')]
+boxes = [s + t for s in rows for t in cols]
+row_units = [[s + t for s in r for t in cols] for r in rows]
+column_units = [[s + t for s in rows for t in c] for c in cols]
+square_units = [[s + t for s in rs for t in cs] for
+                rs in ('ABC', 'DEF', 'GHI') for
+                cs in ('123', '456', '789')]
 unitlist = row_units + column_units + square_units
 units = dict((s, [u for u in unitlist if s in u]) for s in boxes)
-peers = dict((s, set(sum(units[s],[]))-set([s])) for s in boxes)
+peers = dict((s, set(sum(units[s], []))-set([s])) for s in boxes)
+
 
 def single_position(values):
     """
-    Go through all the boxes, and whenever there is a box with a value, eliminate this value from the values of all its peers.
+    Go through all the boxes, and whenever there is a box with a value,
+    eliminate this value from the values of all its peers.
     Input: A sudoku in dictionary form.
     Output: The resulting sudoku in dictionary form.
     """
@@ -30,9 +34,11 @@ def single_position(values):
             values[peer] = values[peer].replace(digit, "")
     return values
 
+
 def single_candidate(values):
     """
-    Go through all the units, and whenever there is a unit with a value that only fits in one box, assign the value to this box.
+    Go through all the units, and whenever there is a unit with a value that
+    only fits in one box, assign the value to this box.
     Input: A sudoku in dictionary form.
     Output: The resulting sudoku in dictionary form.
     """
@@ -43,6 +49,7 @@ def single_candidate(values):
                 values[dplaces[0]] = digit
     return values
 
+
 def naked_twins(values):
     """Eliminate values using the naked twins strategy.
     Args:
@@ -50,63 +57,76 @@ def naked_twins(values):
     Returns:
         the values dictionary with the naked twins eliminated from peers.
     """
-        
     for unit in unitlist:
-        #1. Find twins 
-        twins = [v for v in [values[box] for box in unit] if [values[box] for box in unit].count(v) == 2 and len(v) == 2]
-            
+        # 1. Find twins
+        twins = [v for v in [values[box] for box in unit] if
+                 [values[box] for box in unit].count(v) == 2 and len(v) == 2]
         for box in unit:
             if values[box] in twins:
                 continue
-            for twin in twins :
+            for twin in twins:
                 for digit in twin:
                     values[box] = values[box].replace(digit, "")
     return values
 
+
 def naked_triple(values):
-    values_triples = [a for a,b in Counter([v for k,v in values.items() if len(v)==3]).items() if b>2]
-    triples= [([k for k,v in values.items() if v == value_triple]) for value_triple in values_triples]
+    values_triples = [a for a, b in Counter([v for k, v in values.items() if
+                      len(v) == 3]).items() if b > 2]
+    triples = [([k for k, v in values.items() if v == value_triple]) for
+               value_triple in values_triples]
     for triple in triples:
         for unit in unitlist:
-            if(set(triple).issubset(set(unit))): 
-                values_remove= [x for x in unit if x not in triple]
+            if(set(triple).issubset(set(unit))):
+                values_remove = [x for x in unit if x not in triple]
                 digits = values[triple[0]]
-                for value_remove in values_remove: 
+                for vr in values_remove:
                     for digit in digits:
-                        values[value_remove] = values[value_remove].replace(digit, "")
+                        values[vr] = values[vr].replace(digit, "")
     return values
-            
+
+
 def reduce_puzzle(values):
     """
-    Iterate eliminate() and only_choice(). If at some point, there is a box with no available values, return False.
+    Iterate eliminate() and only_choice(). If at some point,
+    there is a box with no available values, return False.
     If the sudoku is solved, return the sudoku.
-    If after an iteration of both functions, the sudoku remains the same, return the sudoku.
+    If after an iteration of both functions,
+    the sudoku remains the same, return the sudoku.
     Input: A sudoku in dictionary form.
     Output: The resulting sudoku in dictionary form.
     """
-    solved_values = [box for box in values.keys() if len(values[box]) == 1]
+    solved_values = [box for box in values.keys() if
+                     len(values[box]) == 1]
     stalled = False
     while not stalled:
-        solved_values_before = len([box for box in values.keys() if len(values[box]) == 1])
+        solved_values_before = len([box for box in values.keys() if
+                                    len(values[box]) == 1])
         values = single_position(values)
         values = single_candidate(values)
         values = naked_twins(values)
         values = naked_triple(values)
-        solved_values_after = len([box for box in values.keys() if len(values[box]) == 1])
+        solved_values_after = len([box for box in values.keys() if
+                                   len(values[box]) == 1])
         stalled = solved_values_before == solved_values_after
-        if len([box for box in values.keys() if len(values[box]) == 0]):
+        if len([box for box in values.keys() if
+                len(values[box]) == 0]):
             return False
     return values
+
 
 def search(values):
     "Using depth-first search and propagation, try all possible values."
     # First, reduce the puzzle using the previous function
     values = reduce_puzzle(values)
     if values is False:
-        return False ## Failed earlier
-    if all(len(values[s]) == 1 for s in boxes): 
-        return values ## Solved!
-    n,s = min((len(values[s]), s) for s in boxes if len(values[s]) > 1)
+        # Failed earlier
+        return False
+    if all(len(values[s]) == 1 for s in boxes):
+        # Solved!
+        return values
+    n, s = min((len(values[s]), s) for s in boxes if
+               len(values[s]) > 1)
     for value in values[s]:
         new_sudoku = values.copy()
         new_sudoku[s] = value
@@ -114,75 +134,83 @@ def search(values):
         if attempt:
             return attempt
 
+
 def display(values):
     """
     Display the values as a 2-D grid.
     Input: The sudoku in dictionary form
     Output: None
     """
-    width = 1+max(len(values[s]) for s in boxes)
-    line = '+'.join(['-'*(width*3)]*3)
+    width = 1 + max(len(values[s]) for s in boxes)
+    line = '+'.join(['-' * (width * 3)] * 3)
     for r in rows:
-        print(''.join(values[r+c].center(width)+('|' if c in '36' else '')
+        print(''.join(values[r+c].center(width) +
+              ('|' if c in '36' else '')
                       for c in cols))
-        if r in 'CF': print(line)       
+        if r in 'CF':
+            print(line)
     print()
 
+
 def validator(grid):
-    valuesv = dict(zip(boxes,["." if element == "." else element for element in grid]))  
-    answ=[]
-    values_grid = dict(filter(lambda n: n[1]!=".", valuesv.items()))
-    a=[(valuesv[n],[valuesv[p] for p in peers[n]],n) for n in list(values_grid.keys())]
+    valuesv = dict(zip(boxes, ["." if element == "." else
+                   element for element in grid]))
+    answ = []
+    values_grid = dict(filter(lambda n: n[1] != ".", valuesv.items()))
+    a = [(valuesv[n], [valuesv[p] for p in peers[n]], n) for
+         n in list(values_grid.keys())]
     for x in a:
         if x[0] in x[1]:
-            answ.append([False,x[0],x[2]])
-        else: 
+            answ.append([False, x[0], x[2]])
+        else:
             pass
     return answ
+
 
 def transf(values):
     return len("".join([value for value in values.values()]))
 
+
 def tracker(values):
     stalled = False
-    start=0
-    answ=[0,0,0,0,0]
+    start = 0
+    answ = [0, 0, 0, 0, 0]
     while not stalled:
         initial = transf(values)
         values_before = values.values()
         values = single_position(values)
         values_after = transf(values)
-        answ = (answ if values_before == values_after else np.add(answ, [1,0,0,0,0]))
-        
+        answ = (answ if values_before == values_after else
+                np.add(answ, [1, 0, 0, 0, 0]))
         values_before = values_after
         values = single_candidate(values)
         values_after = transf(values)
-        answ = (answ if values_before == values_after else np.add(answ, [0,1,0,0,0])) 
-
+        answ = (answ if values_before == values_after else
+                np.add(answ, [0, 1, 0, 0, 0]))
         values_before = values_after
         values = naked_twins(values)
         values_after = transf(values)
-        answ = (answ if values_before == values_after else np.add(answ, [0,0,1,0,0])) 
-        
+        answ = (answ if values_before == values_after else
+                np.add(answ, [0, 0, 1, 0, 0]))
         values_before = values_after
         values = naked_triple(values)
         values_after = transf(values)
-        answ = (answ if values_before == values_after else np.add(answ, [0,0,0,1,0])) 
-            
-        solved_values = len([box for box in values.keys() if len(values[box]) == 1])
+        answ = (answ if values_before == values_after else
+                np.add(answ, [0, 0, 0, 1, 0]))
+        solved_values = len([box for box in values.keys()
+                             if len(values[box]) == 1])
         if solved_values is 81:
             break
- 
-        stalled = solved_values == 81    
-        
-        if initial == values_after: 
+        stalled = solved_values == 81
+        if initial == values_after:
             start += 1
-            aa = [(len(values[s]), s) for s in boxes if len(values[s]) > 1]
+            aa = [(len(values[s]), s) for
+                  s in boxes if len(values[s]) > 1]
             if len(aa) is 0:
                 pass
             if len(aa) > 0:
-                answ = np.add(answ, [0,0,0,0,1])
-                _,s = min(aa)
+                answ = np.add(answ, [0, 0, 0, 0, 1])
+                _, s = min(aa)
                 for value in values[s]:
                     new_sudoku = values.copy()
                     new_sudoku[s] = value
@@ -193,21 +221,30 @@ def tracker(values):
 
 
 def conv_values(grid):
-     return dict(zip(boxes, ["123456789" if element == "." else element for element in grid]))
-    
+    return dict(zip(boxes, ["123456789" if
+                    element == "." else element for element in grid]))
+
+
 def train_model():
-    df = pd.read_csv('../Omega2020/data/dataset.csv').drop('Unnamed: 0', axis=1)
+    df = pd.read_csv(
+                     '../Omega2020/data/dataset.csv').drop('Unnamed: 0',
+                                                           axis=1)
     df = df.drop(df[df.Level == 'TEST'].index)
     df['Tracker'] = df['Sudoku'].apply(lambda x: tracker(conv_values(x)))
-    df[['Single', 'Candidate', 'Twins', 'Triples', 'Guess']]=  pd.DataFrame(df['Tracker'].values.tolist(), index= df.index)
-    target =['Level']
-    features = ['Single','Candidate','Twins','Triples','Guess']
+    df[['Single', 'Candidate', 'Twins',
+        'Triples', 'Guess']] = pd.DataFrame(df['Tracker'].values.tolist(),
+                                            index=df.index)
+    target = ['Level']
+    features = ['Single', 'Candidate', 'Twins', 'Triples', 'Guess']
     y = df[target]
     X = df[features]
-    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
-    model = LogisticRegression(solver='lbfgs', multi_class='auto',max_iter=1000)
-#Create Pickle
-    outfile = open('difficulty_level_model','wb')
-    pickle.dump(model.fit(X_train, y_train.values.ravel()),outfile)
+    X_train, X_test, y_train, y_test = train_test_split(X, y,
+                                                        test_size=0.2,
+                                                        random_state=42)
+    model = LogisticRegression(solver='lbfgs',
+                               multi_class='auto',
+                               max_iter=1000)
+    outfile = open('difficulty_level_model', 'wb')
+    pickle.dump(model.fit(X_train, y_train.values.ravel()), outfile)
     outfile.close()
-    
+
